test(qa/integration-tests): improve test stability
* merge changes in test classes from develop
